Created Utilities class and namespace
