+++ includes/Utility.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <string>
+
+namespace Util
+{
+	std::string removeTrailingWhitespace(const std::string &str);
+	std::string removeLeadingWhitespace(const std::string &str);
+}
